hosting time

+++ resources/users/routes.py
@@ -1,5 +1,6 @@
 
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_smorest import abort
 from sqlalchemy.exc import IntegrityError
 from schemas import UplateUserSchema, UserSchema, ReviewSchema, AutheuserSchema, UserSchemaNested
@@ -17,14 +18,30 @@ class UserList(MethodView):
         return users
 
    
-
+    @jwt_required()
     @bp.arguments(AutheuserSchema)
     def delete(self, user_data):
-        user= UserModel.query.filter_by(username = user_data['username']).first()
-        if user and user.check_password(user_data['password']):
+        user_id =get_jwt_identity()
+        user = UserModel.query.get(user_id)
+        if user and user.username == user_data['username'] and user.check_password(user_data['password']):
             user.delete()
             return {'message' : f'{user_data["username"]} deleted'},202
         abort(400,message ='username or password invalid')
+
+
+    @jwt_required()
+    @bp.arguments(UplateUserSchema)
+    @bp.response(202, UserSchema)    
+    def put(self,user_data):
+        user_id = get_jwt_identity()
+        user = UserModel.query.get_or_404(user_id, description='user not found')
+        if user and user.check_password(user_data['password']):
+            try:
+                user.from_dict(user_data)
+                user.save()
+                return user
+            except IntegrityError:
+                abort(400, message = 'username or email already taken')
 
 
 @bp.route('/user/<user_id>')
@@ -34,24 +51,15 @@ class User(MethodView):
       return UserModel.query.get_or_404(user_id, description='user not found')
       
     
-    @bp.arguments(UplateUserSchema)
-    @bp.response(202, UserSchema)    
-    def put(self,user_data, user_id):
-        user = UserModel.query.get_or_404(user_id, description='user not found')
-        if user and user.check_password(user_data['password']):
-            try:
-                user.from_dict(user_data)
-                user.save()
-                return user
-            except IntegrityError:
-                abort(400, message = 'username or email already taken')
         
 
-@bp.route('/user/follow/<follower_id>/<followed_id>')
+@bp.route('/user/follow/<followed_id>')
 class FollowUser(MethodView):
     
+    @jwt_required()
     @bp.response(200, UserSchema(many=True))
-    def post(self,follower_id, followed_id):
+    def post(self, followed_id):
+        follower_id = get_jwt_identity()    
         user = UserModel.query.get(follower_id)
         user_to_follow = UserModel.query.get(followed_id)
         if user and user_to_follow:
@@ -59,8 +67,9 @@ class FollowUser(MethodView):
             return user.followed.all()
         abort(400, message = 'invalid user info')
     
-    
-    def put(self,follower_id, followed_id):
+    @jwt_required()
+    def put(self,followed_id):
+        follower_id = get_jwt_identity()
         user = UserModel.query.get(follower_id)
         user_to_unfollow = UserModel.query.get(followed_id)
         if user and user_to_unfollow:
